+++ R/clusterHclust.R
@@ -4,13 +4,15 @@
 #'
 #'
 #' @param dat data to define the mediods from. Samples on rows, features on columns
-#' @param cl A vector with cluster assignments to compare to clRef.``-1'' indicates the sample was not assigned to a cluster.
+#' @param cl A numeric vector with cluster assignments to compare to clRef.``-1'' indicates the sample was not assigned to a cluster.
 #' @param full Logical as to whether to expand to include all samples as tips (useful for heatmap), or to have dendrogram just of the clusters (useful for getBestGenes)
 #' @param unassigned what to do with with unassigned ("-1") cluster; only relevant if full=TRUE. Default is to remove those samples, so that only get heatmap of clustered observations. ``outgroup'' puts the unassigned into an outgroup on the dendrogram. ``cluster'' clusters them with the mediods. 
 #' @param ... arguments passed to hclust
 #' @details If full=TRUE, then the expanded dendrogram is created by giving all of the samples the mediod of the cluster and applying hclust to it; if further unassigned=="cluster", then the expanded dendrogram is created by hclust of the expanded mediod data plus the original unclustered observations.
 #'
 #'
+#' @return A dendrogram for the clusters. If full=TRUE, the number of leaves is equal to the number of samples. If full=FALSE, the number of leaves is equal to the number of clusters.
+
 #' @examples
 #' data(simData)
 #' #create a clustering, for 8 clusters (truth was 3)
@@ -29,6 +31,9 @@ clusterHclust<-function(dat,cl,full=TRUE,unassigned=c("remove","outgroup","clust
 	unassigned<-match.arg(unassigned)
 	if(nrow(dat)!=length(cl)) stop("cl must be the same length as the number of rows of dat")
 	if(is.null(rownames(dat))) rownames(dat)<-as.character(1:nrow(dat))
+	if(is.factor(cl)){warning("cl is a factor. Converting to numeric, which may not result in valid conversion")
+			cl<-as.numeric(as.character(cl))
+	}
 	whRm<- which(cl!=-1)
 	clFactor<-factor(cl[whRm])
 	mediods<-do.call("rbind",by(dat[whRm,],clFactor,function(x){apply(x,2,median)}))
@@ -86,15 +91,25 @@ clusterHclust<-function(dat,cl,full=TRUE,unassigned=c("remove","outgroup","clust
 #'
 #'
 #' @param dat data to perform the test on. Samples on rows, features on columns
-#' @param cl A vector with cluster assignments to compare to clRef.``-1'' indicates the sample was not assigned to a cluster.
+#' @param cl A numeric vector with cluster assignments to compare to clRef.``-1'' indicates the sample was not assigned to a cluster.
+#' @param dendro If provided, is dendrogram providing hierarchical clustering of clusters in cl; mainly useful to speed up calculations if clusterHclust has already been called on the (cl,dat) pair. The default (NULL) is to recalculate it with the given (cl,dat) pair.
 #' @param mergeMethod method for calculating proportion of non-null that will be used to merge clusters (if 'none', no merging will be done). See details for description of methods. 
 #' @param cutoff cutoff for merging clusters based on the proportion of non-nulls in the comparison of the clusters (i.e. value between 0,1, where lower values will make it harder to merge clusters).
 #' @param plotType what type of plotting of dendrogram. If 'all', then all the estimates of proportion non-null will be plotted; if 'mergeMethod', then only the value used in the merging is plotted for each node. 
+#' @param countData logical as to whether input data is a count matrix (in which case log(count+1) will be used for \code{\link{clusterHclust}} and voom correction will be used in \code{\link{getBestGenes}}).
 #' @param ... arguments passed to the \code{\link{plot.phylo}} function of \code{ade4} that plots the dendrogram. 
 #' 
 #' @details "JC" refers to the method of Ji and Cai (2007), and implementation of "JC" method is copied from code available on Jiashin Ji's website, December 16, 2015 (http://www.stat.cmu.edu/~jiashun/Research/software/NullandProp/). "locfdr" refers to the method of Efron (2004) and is implemented in the package \code{\link{locfdr}}. "MB" refers to the method of Meinshausen and Buhlmann (2005) and is implemented in the package \code{\link{howmany}}. "adjP" refers to the proportion of genes that are found significant based on a FDR adjusted p-values (method "BH") and a cutoff of 0.05. 
 #'
 #' @details If \code{mergeMethod} is not equal to 'none' then the plotting will indicate the merged clusters (assuming \code{plotType} is not 'none'). 
+#' @return Returns (invisibly) a list with elements
+#' \itemize{
+
+#' \item{\code{clustering}}{a vector of length equal to nrows(dat) giving the integer-valued cluster ids for each sample. "-1" indicates the sample was not clustered.}
+#' \item{\code{oldClToNew}}{A table of the old cluster labels to the new cluster labels}
+#' \item{\code{propDE}}{A table of the proportions that are DE on each node}
+#' \item{\code{originalClusterDendro}}{The dendrogram on which the merging was based (based on the original clustering)}
+#' }
 #' @examples
 #' data(simData)
 #'  #create a clustering, for 8 clusters (truth was 3)
@@ -102,20 +117,33 @@ clusterHclust<-function(dat,cl,full=TRUE,unassigned=c("remove","outgroup","clust
 #'  sequential=FALSE, clusterDArgs=list(k=8))$cl
 #' #merge clusters with plotting. Note argument 'use.edge.length' can improve readability
 #' mergeResults<-mergeClusters(simData,cl=cl,plot=TRUE,plotType="all",
-#' mergeMethod="adjP",use.edge.length=FALSE)
+#' mergeMethod="adjP",use.edge.length=FALSE,countData=FALSE)
 #' #compare merged to original on heatmap
 #' hclData<-clusterHclust(dat=simData,cl,full=TRUE)
 #' dualHeatmap(cl,heatData=simCount,clusterData=hclData,colorScale=seqPal5,
 #'	annCol=data.frame(Original=cl,Merged=mergeResults$cl,Truth=trueCluster))
 
  
-mergeClusters<-function(dat,cl,mergeMethod=c("none","adjP","locfdr","MB","JC"),cutoff=0.1,plotType=c("none","all","mergeMethod"),...){
-	dendro<-clusterHclust(dat=dat,cl,full=FALSE)
+mergeClusters<-function(dat,cl,dendro=NULL,mergeMethod=c("none","adjP","locfdr","MB","JC"),cutoff=0.1,plotType=c("none","all","mergeMethod"),countData=TRUE,...){
+	if(is.factor(cl)){warning("cl is a factor. Converting to numeric, which may not result in valid conversion")
+			cl<-as.numeric(as.character(cl))
+	}
+	if(!is.null(dendro)){
+		#check valid
+		ncluster<-length(table(cl[cl>0]))
+		if(nobs(dendro)!=ncluster){
+			warning("Not a valid input dendrogram (not equal to the number of non -1 clusters in cl). Will recompute dendrogram")
+			dendro<-NULL
+		}
+	}
+	if(is.null(dendro)){
+		dendro<-if(countData) clusterHclust(dat=log(dat+1),cl,full=FALSE) else clusterHclust(dat=dat,cl,full=FALSE)		
+	}
   	mergeMethod<-match.arg(mergeMethod)
   	plotType<-match.arg(plotType)
 	if(plotType=="mergeValue" & mergeMethod=="none") stop("can only plot merge method values if one method is selected")
 	#get test-statistics for the contrasts corresponding to each node (and return all)
-	sigTable<-getBestGenes(cl,dat,type=c("Dendro"),dendro=dendro,returnType=c("Table"),contrastAdj=c("All"),number=ncol(dat),p.value=1)
+	sigTable<-getBestGenes(cl,dat,type=c("Dendro"),dendro=dendro,returnType=c("Table"),contrastAdj=c("All"),number=ncol(dat),p.value=1,voomCorrection=countData)
 	#divide table into each node.
 	sigByNode<-by(sigTable,sigTable$ContrastName,function(x){
 		mb<-.myTryFunc(pvalues=x$P.Value,FUN=.m1_MB)
@@ -187,7 +215,7 @@ mergeClusters<-function(dat,cl,mergeMethod=c("none","adjP","locfdr","MB","JC"),c
 	nodePropTable<-do.call("rbind",sigByNode)
 	nodePropTable<-data.frame("Node"=names(sigByNode),"Contrast"=sigTable$Contrast[match(names(sigByNode),sigTable$ContrastName)],nodePropTable)
 
-	invisible(list(cl=newcl,oldClToNew=table(Original=cl,New=newcl),propDE=nodePropTable,originalClusterDendro=dendro))
+	invisible(list(clustering=newcl,oldClToNew=table(Original=cl,New=newcl),propDE=nodePropTable,originalClusterDendro=dendro))
 }
 
 .myTryFunc<-function(FUN,...){